+++ packages/eth-extended/CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 - **Breaking change** Removed deprecated `module://aleth.io/block/details` & `adapter://aleth.io/block/details`
 - **Breaking change** Renamed `adapter://aleth.io/extended/block/details` to `adapter://aleth.io/full/block/details`
+- **Breaking change** Renamed `adapter://aleth.io/extended/uncle/details` to `adapter://aleth.io/full/uncle/details`
 
 # v4.0.3